+++ test/storages/sequelize.test.js
@@ -1,6 +1,6 @@
 const { expect } = require('chai');
 const helper = require('../helper');
-const Storage = require('../../src/storages/SequelizeStorage');
+const { SequelizeStorage: Storage } = require('../../lib/storages/SequelizeStorage');
 const Sequelize = require('sequelize');
 const { join } = require('path');
 const { v4: uuid } = require('uuid');
